Fix null asset bug
When there is not an asset.json filled out for the pool the asset would be null and cause the transfer to not get posted due to a destructoring error. This fixes that bug and also adds a note on the post to update the asset.json for the pool.